add mixdown function, only evaluate node if any inputs are active, some other performance enhancements

+++ editor/model/connection/ExtractConnectionSink.h
@@ -8,7 +8,7 @@ namespace AxiomModel {
         Q_OBJECT
 
     public:
-        using ActiveSlotFlags = uint16_t;
+        using ActiveSlotFlags = uint32_t;
 
         explicit ExtractConnectionSink(MaximRuntime::Control *runtime, Type baseType);
